feat: integrate Iyzico payment gateway

+++ Controllers/OrderController.cs
@@ -1,6 +1,9 @@
 using dotnet_store.Data;
 using dotnet_store.Models;
 using dotnet_store.Services;
+using Iyzipay;
+using Iyzipay.Model;
+using Iyzipay.Request;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
@@ -12,11 +15,13 @@ namespace dotnet_store.Controllers;
 public class OrderController : Controller
 {
     private ICartService _cartService;
+    private readonly IConfiguration _config;
     private readonly DataContext _context;
-    public OrderController(ICartService cartService, DataContext context)
+    public OrderController(ICartService cartService, DataContext context, IConfiguration config)
     {
         _cartService = cartService;
         _context = context;
+        _config = config;
     }
 
     [Authorize(Roles = "Admin")]
@@ -56,7 +61,7 @@ public class OrderController : Controller
             var order = new Order
             {
                 NameSurname = model.NameSurname,
-                PhoneNumber = model.PhoneNumber,
+                PhoneNumber = "+905" + model.PhoneNumber,
                 City = model.City,
                 PostalCode = model.PostalCode,
                 OpenAddress = model.OpenAddress,
@@ -64,7 +69,7 @@ public class OrderController : Controller
                 OrderTime = DateTime.Now,
                 TotalAmount = cart.Total(),
                 Username = username,
-                OrderItem = cart.CartItems.Select(cartItem => new OrderItem
+                OrderItem = cart.CartItems.Select(cartItem => new Data.OrderItem
                 {
                     ProductId = cartItem.ProductId,
                     Product = cartItem.Product,
@@ -73,13 +78,17 @@ public class OrderController : Controller
                 }).ToList()
             };
 
-            await _context.Orders.AddAsync(order);
-            _context.Carts.Remove(cart);
-            await _context.SaveChangesAsync();
+            var payment = await ProcessPayment(model, cart);
 
-            return RedirectToAction("Completed", new { orderId = order.Id });
+            if (payment.Status == "success")
+            {
+                await _context.Orders.AddAsync(order);
+                _context.Carts.Remove(cart);
+                await _context.SaveChangesAsync();
+                return RedirectToAction("Completed", new { orderId = order.Id });
+            }
+            ModelState.AddModelError("", payment.ErrorMessage);
         }
-
         ViewBag.Cart = cart;
         return View(model);
     }
@@ -94,5 +103,74 @@ public class OrderController : Controller
         var username = User.Identity?.Name;
         var orders = await _context.Orders.Include(i => i.OrderItem).ThenInclude(i => i.Product).Where(i => i.Username == username).ToListAsync();
         return View(orders);
+    }
+
+    private async Task<Payment> ProcessPayment(OrderCreateModel model, Cart cart)
+    {
+        Options options = new Options();
+        options.ApiKey = _config["PaymentAPI:ApiKey"];
+        options.SecretKey = _config["PaymentAPI:SecretKey"];
+        options.BaseUrl = "https://sandbox-api.iyzipay.com";
+
+        CreatePaymentRequest request = new CreatePaymentRequest();
+        request.Locale = Locale.TR.ToString();
+        request.ConversationId = Guid.NewGuid().ToString();
+        request.Price = (cart.Total() - cart.Discount(1000)).ToString(System.Globalization.CultureInfo.InvariantCulture);
+        request.PaidPrice = (cart.Total() - cart.Discount(1000)).ToString(System.Globalization.CultureInfo.InvariantCulture);
+        request.Currency = Currency.TRY.ToString();
+        request.Installment = 1;
+        request.BasketId = "B67832";
+        request.PaymentChannel = PaymentChannel.WEB.ToString();
+        request.PaymentGroup = PaymentGroup.PRODUCT.ToString();
+
+        PaymentCard paymentCard = new PaymentCard();
+        paymentCard.CardHolderName = model.CartOwner;
+        paymentCard.CardNumber = model.CartNumber;
+        paymentCard.ExpireMonth = model.CartExpirationMonth;
+        paymentCard.ExpireYear = model.CartExpirationYear;
+        paymentCard.Cvc = model.CartCVV;
+        paymentCard.RegisterCard = 0;
+        request.PaymentCard = paymentCard;
+
+        Buyer buyer = new Buyer();
+        buyer.Id = cart.CustomerId;
+        buyer.Name = model.NameSurname;
+        buyer.Surname = "Doe";
+        buyer.GsmNumber = model.PhoneNumber;
+        buyer.Email = "email@email.com";
+        buyer.IdentityNumber = "74300864791";
+        buyer.LastLoginDate = "2015-10-05 12:43:35";
+        buyer.RegistrationDate = "2013-04-21 15:12:09";
+        buyer.RegistrationAddress = "Nidakule Göztepe, Merdivenköy Mah. Bora Sok. No:1";
+        buyer.Ip = "85.34.78.112";
+        buyer.City = model.City;
+        buyer.Country = "Turkey";
+        buyer.ZipCode = model.PostalCode;
+        request.Buyer = buyer;
+
+        Iyzipay.Model.Address address = new Iyzipay.Model.Address();
+        address.ContactName = model.NameSurname;
+        address.City = model.City;
+        address.Country = "Turkey";
+        address.Description = model.OpenAddress;
+        address.ZipCode = model.PostalCode;
+        request.ShippingAddress = address;
+        request.BillingAddress = address;
+
+        List<BasketItem> basketItems = new List<BasketItem>();
+        foreach (var item in cart.CartItems)
+        {
+            BasketItem basketItem = new BasketItem();
+            basketItem.Id = item.CartId.ToString();
+            basketItem.Name = item.Product.Name;
+            basketItem.Category1 = item.Product.Category.Name ?? "General";
+            basketItem.ItemType = BasketItemType.PHYSICAL.ToString();
+            basketItem.Price = (item.Product.Price + item.Product.Price * 0.2 - 1000).ToString(System.Globalization.CultureInfo.InvariantCulture);
+
+            basketItems.Add(basketItem);
+        }
+
+        request.BasketItems = basketItems;
+        return await Payment.Create(request, options);
     }
 }
